// frontend/src/lib/backend/apiV1/schema.d.ts
export interface components {
    schemas: {
        ApiResponseString: {
            success?: boolean;
            message?: string;
            content?: string;
        };
        ProductRequest: {
            productName: string;
            category: string;
            /** Format: int32 */
            price?: number;
            imageUrl?: string;
        };
        ApiResponseOrderResponseDTO: {
            success?: boolean;
            message?: string;
            content?: components["schemas"]["OrderResponseDTO"];
        };
        OrderItemResponseDTO: {
            /** Format: int64 */
            id?: number;
            productName?: string;
            /** Format: int32 */
            count?: number;
            /** Format: int32 */
            price?: number;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            modifiedAt?: string;
            /** Format: int32 */
            totalPrice?: number;
        };
        OrderResponseDTO: {
            /** Format: int64 */
            id?: number;
            email?: string;
            status?: string;
            /** Format: int32 */
            totalPrice?: number;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            modifiedAt?: string;
            orderItems?: components["schemas"]["OrderItemResponseDTO"][];
            empty?: boolean;
        };
        Order: {
            /** Format: int64 */
            readonly id?: number;
            /** Format: date-time */
            readonly createDate?: string;
            /** Format: date-time */
            readonly modifyDate?: string;
            /**
             * @description 이메일
             * @example test@gmail.com
             */
            email: string;
            /**
             * @description 상품 상태, 기본 값으로 PAYMENT_COMPLETED 입니다.
             * @example PAYMENT_COMPLETED
             * @enum {string}
             */
            readonly status: "UNKNOWN" | "CANCELLED" | "PAYMENT_COMPLETED" | "PREPARING" | "SHIPPING" | "COMPLETED";
            /**
             * Format: int32
             * @description 주문 금액
             * @example 1000
             */
            totalPrice: number;
            /** @description 주문에 포함된 상품 목록 */
            orderItems?: components["schemas"]["OrderItem"][];
        };
        OrderItem: {
            /** Format: int64 */
            readonly id?: number;
            /** Format: date-time */
            readonly createDate?: string;
            /** Format: date-time */
            readonly modifyDate?: string;
            /**
             * @description 상품 이름
             * @example 원두 1
             */
            productName: string;
            /**
             * Format: int32
             * @description 상품 수량 (1 이상)
             * @example 1
             */
            count: number;
            /**
             * Format: int32
             * @description 상품 가격 (1 이상)
             * @example 1000
             */
            price: number;
        };
        MemberLoginReqBody: {
            username: string;
            password: string;
        };
        ApiResponseMemberLoginResBody: {
            success?: boolean;
            message?: string;
            content?: components["schemas"]["MemberLoginResBody"];
        };
        MemberLoginResBody: {
            apiKey: string;
            accessToken: string;
        };
        ApiResponsePageDtoProductDto: {
            success?: boolean;
            message?: string;
            content?: components["schemas"]["PageDtoProductDto"];
        };
        PageDtoProductDto: {
            /** Format: int64 */
            totalItems: number;
            items: components["schemas"]["ProductDto"][];
            /** Format: int64 */
            totalPages: number;
            /** Format: int32 */
            currentPageNumber: number;
            /** Format: int32 */
            pageSize: number;
        };
        ProductDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createDate: string;
            /** Format: date-time */
            modifyDate: string;
            productName: string;
            /** Format: int32 */
            price: number;
            imageUrl?: string;
            category: string;
        };
        ApiResponseProductDto: {
            success?: boolean;
            message?: string;
            content?: components["schemas"]["ProductDto"];
        };
        ApiResponseListProductDto: {
            success?: boolean;
            message?: string;
            content?: components["schemas"]["ProductDto"][];
        };
        ApiResponseListOrderResponseDTO: {
            success?: boolean;
            message?: string;
            content?: components["schemas"]["OrderResponseDTO"][];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}